% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsFunc-click.R
\name{click}
\alias{click}
\title{Click on a Shiny button}
\usage{
click(id)
}
\arguments{
\item{id}{The id of the button}
}
\description{
The \code{click()} function can be used to programatically simulate a click
on a Shiny \code{actionButton()}.
}
\note{
\code{shinyjs} must be initialized with a call to \code{useShinyjs()}
in the app's ui.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useShinyjs(),  # Set up shinyjs
      actionButton("btn", "Click me"),actionButton("gg","or me")
    ),
    server = function(input, output) {
    observeEvent(input$gg, click('btn'))
      observeEvent(input$btn, {
        cat('clicked! ',input$btn)
      })
    }
  )
}
}
\seealso{
\code{\link[shinyjs]{useShinyjs}},
\code{\link[shinyjs]{runExample}}
}
