% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourPickerGadget.R
\name{colourPicker}
\alias{colourPicker}
\title{Colour picker gadget}
\usage{
colourPicker()
}
\value{
Vector of selected colours
}
\description{
This gadget lets you choose colours easily. You can select multiple colours,
and you can either choose any RGB colour, or browse through R colours.
}
\note{
This gadget returns a vector of colours that can be assigned to a variable.
If instead you want to get a text representation of the colours that can
embedded into code, use the addin from the RStudio Addins menu.
}
\examples{
if (interactive()) {
  cols <- colourPicker()
}
}

