% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hidden.R
\name{hidden}
\alias{hidden}
\title{Initialize a Shiny tag as hidden}
\usage{
hidden(tag)
}
\arguments{
\item{tag}{Shiny tag}
}
\value{
The tag that was passed in as a parameter hidden
}
\description{
Create a Shiny tag that is invisible when the app starts.
}

