% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/onclick.R
\name{onclick}
\alias{onclick}
\title{Add an onclick event handler}
\usage{
onclick(id, expr, add = FALSE)
}
\arguments{
\item{id}{The id of the HTML element}

\item{expr}{The R expression to run after the element gets clicked}

\item{add}{If TRUE, add expr to be executed after any previously set onclick
handlers; otherwise (the default) expr will overwrite any previous onclick
expressions}
}
\description{
Run an R expression when an HTML element is clicked
}

