% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound.R
\name{playSound}
\alias{playSound}
\title{Play a sound}
\usage{
playSound(src, mime = NULL, file, immediate = TRUE)
}
\arguments{
\item{src}{A relative or absolute URL. Mutually exclusive with \code{file}.}

\item{mime}{The mime type of the audio data, or \code{NULL} to guess based on
file extension.}

\item{file}{A path to a file on disk. Mutually exclusive with \code{src}.}

\item{immediate}{Set to \code{FALSE} to delay sending the sound command to 
the browser until all Shiny outputs have finished calculating and all
observers have been run.}
}
\description{
Play a sound in the browser, using either a (relative or absolute) source URL
or a file on disk. Audio format support is browser-dependent (at the time of 
this writing, MP3 and AAC in MP4 are well supported, but see 
\href{https://developer.mozilla.org/en-US/docs/Web/HTML/Supported_media_formats#Browser_compatibility}{MDN}
for current information). Mobile browsers are not supported, as they 
generally do not allow audio to be automatically played without a user 
gesture.
}
\examples{
if (interactive()) {
  library(shiny)
  
  shinyApp(
    ui = fluidPage(
      useShinyjs(),
      actionButton("play", "Play sound")
    ),
    server = function(input, output) {
      observeEvent(input$play, {
        playSound(file = system.file("mp3/chime.mp3", package="shinyjs"))
      })
    }
  )
}
  
}

