% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{Refresh the page}
\usage{
refresh()
}
\description{
Refresh the page
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    useShinyjs(),
    textInput("text", "Text", "text"),
    actionButton("refresh", "Refresh")
  )

  server <- function(input, output, session) {
    observeEvent(input$refresh, {
      refresh()
    })
  }

 shinyApp(ui, server)
}

}
