% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcode.R
\name{runcode}
\alias{runcode}
\alias{runcodeUI}
\alias{runcodeServer}
\title{Construct to let you run arbitrary R code live in a Shiny app}
\usage{
runcodeUI(code = "", type = c("text", "textarea", "ace"), width = NULL,
  height = NULL, includeShinyjs = NULL)

runcodeServer(show_output = FALSE, preprocessor = NULL)
}
\arguments{
\item{code}{The initial R code to show in the text input when the app loads}

\item{type}{One of \code{"text"} (default), \code{"textarea"}, or \code{"ace"}.
When using a text input, the R code will be limited to be typed within a single line,
and is the recommended option. Textarea should be used if you want to write
long multi-line R code. Note that you can run multiple expressions even in
a single line by appending each R expression with a semicolon.
Use of the \code{"ace"} option requires the \code{shinyAce} package.}

\item{width}{The width of the editable code input (ignored when
\code{type="ace"})}

\item{height}{The height of the editable code input (ignored when
\code{type="text"})}

\item{includeShinyjs}{Deprecated. You should always make sure to initialize
shinyjs using \code{\link[shinyjs]{useShinyjs}}.}

\item{show_output}{If \code{TRUE} the results of executing the R code will be
placed into \code{output$runcode_output} using \code{shiny::renderText} to
allow display via adding \code{textOutput("runcode_output", container=pre)}.}

\item{preprocessor}{Function accepting a single vector argument, which will
be called on the user-provided R code before evaluating it.}
}
\description{
Sometimes when developing a Shiny app, it's useful to be able to run some R
code on-demand. This construct provides your app with a text input where you
can enter any R code and run it immediately.\cr\cr
This can be useful for testing
and while developing an app locally, but it \strong{should not be included in
an app that is accessible to other people}, as letting others run arbitrary R
code can open you up to security attacks.\cr\cr
To use this construct, you must add a call to \code{runcodeUI()} in the UI
of your app, and a call to \code{runcodeServer()} in the server function. You
also need to initialize shinyjs with a call to \code{useShinyjs()} in the UI.
}
\details{
If a function is provided to the \code{preprocss} argument of
\code{runcodeServer}, this will be called on the user-provided R code before
evaluating it.  This allows, for instance, conversion of non-ASCII open and
closing quote characters (unhelpfully) created by some browsers to standard
ASCII quote characters (see the second example).

If \code{show_output=TRUE}, the results of executing the R code will be
placed into \code{output$runcode_output} using \code{shiny::renderText} to
allow you to display this, if desired, by adding
\code{textOutput("runcode_output", container=pre)} to the UI of your
app (see the second example).
}
\note{
You can only have one \code{runcode} construct in your shiny app.
Calling this function multiple times within the same app will result in
unpredictable behaviour.
}
\examples{

# One-line of code, no display of output
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useShinyjs(),  # Set up shinyjs
      runcodeUI(code = "shinyjs::alert('Hello!')")
    ),
    server = function(input, output) {
      runcodeServer()
    }
  )
}

# Muti-line code - pre-process input to convert non-ASCII character to ASCII
# and then display output as if entered on the console
if (interactive()) {
  library(shiny)
  library(stringi)

  to_ascii <- function(x) stri_trans_general(x, "Any-Latin; Latin-ASCII")

  shinyApp(
    ui = fluidPage(
      shiny::tags$label("R code to execute:"),
      useShinyjs(),  # Set up shinyjs
      runcodeUI(code = "shinyjs::alert('Hello!')\\nx <- rnorm(10)\\nstem(x)",
                type = "textarea",
                height="10em",
                width="80em"),
      shiny::tags$br(),
      shiny::tags$label("Output:"),
      textOutput("runcode_output", container=pre)
    ),
    server = function(input, output) {
      runcodeServer(show_output=TRUE, preprocess=to_ascii)
    }
  )
}
}
\seealso{
\code{\link[shinyjs]{useShinyjs}}
}
