% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcode.R
\name{runcodeUI}
\alias{runcodeUI}
\title{Deprecated - use \code{runcode()} instead}
\usage{
runcodeUI(...)
}
\arguments{
\item{...}{All parameters acceptd by \code{\link[=runcode]{runcode()}}}
}
\description{
Deprecated - use \code{runcode()} instead
}
