% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run shinyjs examples}
\usage{
runExample(example = "sandbox")
}
\arguments{
\item{example}{The add to launch}
}
\description{
Launch a shinyjs example Shiny app that allows you to see how to use easily
use shinyjs in an app.
}

