% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-colourInput.R, R/shinyjs-package.R
\name{colourInput}
\alias{colourInput}
\alias{updateColourInput}
\alias{colourPicker}
\alias{shinyjs-defunct}
\title{Defunct functions in shinyjs}
\usage{
colourInput(...)

updateColourInput(...)

colourPicker(...)
}
\description{
\itemize{
\item \strong{colourInput()} Moved to the \{colourpicker\} package
\item \strong{updateColourInput()} Moved to the \{colourpicker\} package
\item \strong{colourPicker()} Moved to the \{colourpicker\} package
\item \strong{runExample()} Use \code{shiny::runExample(package = "shinyjs")}
}
}
\keyword{internal}
