% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyjs-package.R
\docType{package}
\name{shinyjs-package}
\alias{shinyjs}
\alias{shinyjs-package}
\title{shinyjs: Easily Improve the User Experience of Your Shiny Apps in Seconds}
\description{
Perform common useful JavaScript operations in Shiny apps that will greatly improve your apps without having to know any JavaScript. Examples include: hiding an element, disabling an input, resetting an input back to its original value, delaying code execution by a few seconds, and many more useful functions for both the end user and the developer. 'shinyjs' can also be used to easily call your own custom JavaScript functions from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://deanattali.com/shinyjs/}
  \item \url{https://github.com/daattali/shinyjs}
  \item Report bugs at \url{https://github.com/daattali/shinyjs/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID})

}
\keyword{internal}
