% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyjs.R
\docType{package}
\name{shinyjs}
\alias{shinyjs}
\alias{shinyjs-package}
\title{shinyjs}
\description{
Perform common JavaScript operations in Shiny apps using plain R code.
}
\details{
\code{shinyjs} lets you perform common JavaScript operations in Shiny
applications without having to know any JavaScript. Examples include: hiding an
element, disabling an input, resetting an input back to its original value,
delaying code execution by a few seconds, and many more useful functions.
\code{shinyjs} also includes a colour picker widget, a colour picker RStudio
addin, and can also be used to easily run your own custom JavaScript functions
from R.

View a \href{http://daattali.com/shiny/shinyjs-demo/}{demo Shiny app}
or see the full \href{https://github.com/daattali/shinyjs}{README} on GitHub.
}

