% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jsFuncs.R
\name{stateFuncs}
\alias{disable}
\alias{enable}
\alias{stateFuncs}
\alias{toggleState}
\title{Enable/disable an input element}
\usage{
enable(...)

disable(...)

toggleState(...)
}
\description{
Enable/disable an input element
}

