% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/useShinyjs.R
\name{useShinyjs}
\alias{useShinyjs}
\title{Use shinyjs}
\usage{
useShinyjs()
}
\value{
Scripts that shinyjs requires that are automatically inserted
to the app's head tag
}
\description{
Set up a Shiny app to use shinyjs. This function needs to be called from a
Shiny app's ui.
}

