% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jsFuncs.R
\name{visibilityFuncs}
\alias{hide}
\alias{show}
\alias{toggle}
\alias{visibilityFuncs}
\title{Display/hide an element}
\usage{
show(id, anim = FALSE, animType = "slide", time = 0.5)

hide(id, anim = FALSE, animType = "slide", time = 0.5)

toggle(id, anim = FALSE, animType = "slide", time = 0.5)

show(...)

hide(...)

toggle(...)
}
\arguments{
\item{id}{The id of the Shiny tag}

\item{anim}{if TRUE then animate the behaviour}

\item{animType}{The type of animation to use, one of 'slide' or 'fade'}

\item{time}{The number of seconds to make the animation last.}
}
\description{
Display or hide an HTML element.
}
\details{
show makes an element visible, hide makes an element invisible, toggle
displays the element if it it hidden and hides it if it is visible.
}

