% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jsFuncs.R
\name{classFuncs}
\alias{addClass}
\alias{classFuncs}
\alias{removeClass}
\alias{toggleClass}
\title{Add/remove CSS class}
\usage{
addClass(...)

removeClass(...)

toggleClass(...)
}
\arguments{
\item{...}{The following parameters are available:
  \describe{
    \item{\strong{id}}{id of the Shiny tag}
    \item{class}{CSS class}
  }
\tabular{ll}{
  \code{\strong{first}}         \tab if of the shiny tag \cr
  \code{\strong{second item}}   \tab asdfdsaf dsadsa fdsa fasfa \cr
   }
  \itemize{
    \item \strong{First} item
    \item Second item
  }}

\item{class}{The CSS class}
}
\description{
Add or remove a CSS class from an HTML element.
}
\details{
addClass adds a CSS class, removeClass removes a cssClass, toggleClass
adds the class if it is not set and removes the class if it is already
set.
}

