% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourPickerGadget.R
\name{colourPicker}
\alias{colourPicker}
\title{This is a gadget, you can have in your script: `cols <- colourPicker(3)`}
\usage{
colourPicker(numCols = 1)
}
\description{
This is a gadget, you can have in your script: `cols <- colourPicker(3)`
}

