% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hidden.R
\name{hidden}
\alias{hidden}
\title{Initialize a Shiny tag as hidden}
\usage{
hidden(tag)
}
\arguments{
\item{tag}{Shiny tag to make invisible}
}
\value{
The tag that was given as an argument in a hidden state.
}
\description{
Create a Shiny tag that is invisible when the Shiny app starts. The tag can
be made visible later with \code{shinyjs::toggle} or \code{shinyjs::show}.
}
\note{
\code{shinyjs} must be initialized with a call to \code{useShinyjs()}
in the app's ui.
}
\examples{
if (interactive()) {
  shiny::shinyApp(
    ui = shiny::fluidPage(
      useShinyjs(),  # Set up shinyjs
      shiny::actionButton("btn", "Click me"),
      hidden(
        shiny::p(id = "element", "I was born invisible")
      )
    ),
    server = function(input, output, session) {
      shiny::observe({
        if (input$btn == 0) {
          return(NULL)
        }
        show("element")
      })
    }
  )
}
}
\seealso{
\code{\link[shinyjs]{useShinyjs}},
\code{\link[shinyjs]{toggle}},
\code{\link[shinyjs]{show}},
\code{\link[shinyjs]{hide}}
}

