% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extendShinyjs.R
\docType{data}
\name{js}
\alias{js}
\title{Call user-defined JavaScript functions from R}
\format{\preformatted{<environment: 0x0000000008d4dc00> 
}}
\usage{
js
}
\description{
Call user-defined JavaScript functions from R
}
\seealso{
\code{\link[shinyjs]{extendShinyjs}}
}
\keyword{internal}

