% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendShinyjs.R
\docType{data}
\name{js}
\alias{js}
\title{Call user-defined JavaScript functions from R}
\format{An object of class \code{environment} of length 0.}
\usage{
js
}
\description{
Call user-defined JavaScript functions from R
}
\seealso{
\code{\link[shinyjs]{extendShinyjs}}
}
\keyword{internal}

