% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run shinyjs examples}
\usage{
runExample(example)
}
\arguments{
\item{example}{The app to launch}
}
\description{
Launch a \code{shinyjs} example Shiny app that shows how to
easily use \code{shinyjs} in an app.
}
\details{
Run without any arguments to see a list of available example apps.
The "demo" example is also
\href{http://daattali.com:3838/shinyjs-demo/}{available online}
to experiment with.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  # List all available example apps
  runExample()

  runExample("sandbox")
  runExample("demo")
}
}

