% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyjs-package.R
\name{shinyjs-defunct}
\alias{shinyjs-defunct}
\title{Defunct functions in shinyjs}
\description{
\itemize{
\item{\bold{colourInput()}} {Moved to the \code{colourpicker} package.}
\item{\bold{updateColourInput()}} {Moved to the \code{colourpicker} package.}
\item{\bold{colourPicker()}} {Moved to the \code{colourpicker} package.}
}
}
