% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-colourInput.R, R/shinyjs.R
\name{colourInput}
\alias{colourInput}
\alias{updateColourInput}
\alias{colourPicker}
\alias{shinyjs-defunct}
\title{Defunct functions in shinyjs}
\usage{
colourInput(...)

updateColourInput(...)

colourPicker(...)
}
\description{
\itemize{
  \item{\bold{colourInput()}} {Moved to the \code{colourpicker} package.}
  \item{\bold{updateColourInput()}} {Moved to the \code{colourpicker} package.}
  \item{\bold{colourPicker()}} {Moved to the \code{colourpicker} package.}
}
}
\keyword{internal}
