% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyjs.R
\docType{package}
\name{shinyjs}
\alias{shinyjs}
\alias{shinyjs-package}
\title{shinys}
\description{
Easily Run Common JavaScript Functions in Shiny using Plain R Code
}
\details{
Using shinyjs lets you bring the power and flexibility of
JavaScript into any Shiny application without requiring any JavaScript
knowledge.  Many useful JavaScript functions are made available by
shinyjs with a simple R interface so that you don't have to write any
JavaScript code. Even if you do know JavaScript, shinyjs can be used as
convenience functions to avoid dealing with message passing and writing
JavaScript code.

View a \href{http://daattali.com:3838/shinyjs-demo/}{demo Shiny app}
or see the \href{https://github.com/daattali/shinyjs}{README} on GitHub.
}

