% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyjs.R
\docType{package}
\name{shinyjs}
\alias{shinyjs}
\alias{shinyjs-package}
\title{shinyjs}
\description{
Perform common JavaScript operations in Shiny apps using plain R code.
}
\details{
\code{shinyjs} lets you perform common JavaScript operations in Shiny
applications without having to know any JavaScript. Many useful JavaScript
functions are made available by shinyjs with a simple R interface so that
you don't have to write any JavaScript code. Even if you do know JavaScript,
shinyjs can be used as convenience functions to avoid dealing with message
passing and writing JavaScript code.

View a \href{http://daattali.com:3838/shinyjs-demo}{demo Shiny app}
or see the full \href{https://github.com/daattali/shinyjs}{README} on GitHub.
}

