% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/useShinyjs.R
\name{useShinyjs}
\alias{useShinyjs}
\title{Use shinyjs}
\usage{
useShinyjs()
}
\value{
Scripts that \code{shinyjs} requires that are automatically inserted
to the app's \code{<head>} tag.
}
\description{
Set up a Shiny application to use \code{shinyjs}. This function needs to be
called from a Shiny app's UI.
}
\examples{
shiny::shinyUI(
  useShinyjs()
)
}
\seealso{
\code{\link[shinyjs]{runExample}}
}

