% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/useShinyjs.R
\name{useShinyjs}
\alias{useShinyjs}
\title{Use shinyjs}
\usage{
useShinyjs()
}
\value{
Scripts that \code{shinyjs} requires that are automatically inserted
to the app's \code{<head>} tag.
}
\description{
Set up a Shiny application to use \code{shinyjs}. This function needs to be
called from a Shiny app's UI. Most \code{shinyjs} functions will not work
if \code{useShinyjs()} is not called first.
}
\examples{
shiny::shinyUI(
  useShinyjs()
)

if (interactive()) {
  shiny::shinyApp(
    ui = shiny::fluidPage(
      useShinyjs(),  # Set up shinyjs
      shiny::actionButton("btn", "Click me"),
      shiny::p(id = "element", "Watch what happens to me")
    ),
    server = function(input, output, session) {
      shiny::observe({
        if (input$btn == 0) {
          return(NULL)
        }
        # Run a simply shinyjs function
        toggle("element")
      })
    }
  )
}
}
\seealso{
\code{\link[shinyjs]{runExample}}
}

