% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jsFuncs.R
\name{visibilityFuncs}
\alias{hide}
\alias{show}
\alias{toggle}
\alias{visibilityFuncs}
\title{Display/hide an element}
\usage{
show(...)

hide(...)

toggle(...)
}
\arguments{
\item{...}{The following parameters are available:
\tabular{ll}{
  \strong{\code{id}}         \tab The id of the Shiny tag \cr
  \strong{\code{anim}}       \tab If TRUE then animate the behaviour \cr
  \strong{\code{animType}}   \tab The type of animation to use,
                                  one of 'slide' or 'fade' \cr
  \strong{\code{time}}       \tab The number of seconds to make the
                                  animation last. \cr
}}
}
\description{
Display or hide an HTML element.
}
\details{
show makes an element visible, hide makes an element invisible, toggle
displays the element if it it hidden and hides it if it is visible.
}
\section{Passing arguments}{

The arguments can be passed either named or unnamed
}

